Undoing code movement
