+++ packages/dds/shared-object-base/src/remoteObjectHandle.ts
@@ -3,16 +3,14 @@
  * Licensed under the MIT License.
  */
 
+import { RuntimeHeaders } from "@fluidframework/container-runtime/internal";
 import { FluidObject, IRequest } from "@fluidframework/core-interfaces";
 import {
 	IFluidHandleContext,
 	type IFluidHandleInternal,
 } from "@fluidframework/core-interfaces/internal";
 import { assert } from "@fluidframework/core-utils/internal";
-
-import { FluidHandleBase } from "./handles.js";
-
-import { responseToException, RuntimeHeaders } from "@fluidframework/runtime-utils/internal";
+import { FluidHandleBase, responseToException } from "@fluidframework/runtime-utils/internal";
 
 /**
  * This handle is used to dynamically load a Fluid object on a remote client and is created on parsing a serialized
@@ -20,8 +18,6 @@ import { responseToException, RuntimeHeaders } from "@fluidframework/runtime-uti
  * This class is used to generate an IFluidHandle when de-serializing any all handles (including handles to DDSes,
  * custom objects) that are stored in SharedObjects. The Data Store or SharedObject corresponding to the
  * IFluidHandle can be retrieved by calling `get` on it.
- *
- * @internal
  */
 export class RemoteFluidObjectHandle extends FluidHandleBase<FluidObject> {
 	public readonly isAttached = true;
